feat(backend): serve only https
...even for development. Creating a local CA authority and certificates
isn't that hard and the reward is less production-only code that could
be untested or could be hiding bugs.

For convenience, I provided 2 scripts, `genRoot.sh` and `genCert.sh`,
that respectively generate a local CA authority (without installing it)
and a key/certificate pair for localhost and `dev.psylean.it`.