feat(backend): implement account registration
This endpoint that accepts POST requests with form data formatted as
`x-www-form-urlencoded`, `form-data` or JSON. These are the necessary
fields:
- `email`: user email
- `password`: user password
- `name`: user real name
- `job`: one value between 'psicologo' and 'psichiatra'
- `region`: lowercase region name in which user is registered as doctor
- `registerNumber`: user's doctor register number
- `_csrf`: content of 'XSRF-TOKEN' cookie

The values received will be validated and they will be saved. If a
validation error occurs, the server will respond with a 400
response containing JSON detailing the errors occured:

    {
      "error": "ValidationError",
      "message": plain text message
      "errors: [
        {
          "path": key that caused the error,
          "kind": type of error (e.g. "required", "regexp"),
          "value": value received
        }...
      ]
    }

A verification email with the verification link
(`/users/verify?q=:code`) will be sent to the provided email address.
Before 24 hours the account must be verified by sending a POST request
to `/api/doctor/verify/:code`. If the code is valid, the `emailVerified`
flag is set and a 200 response is returned, otherwise a 404 one. If the
account is not verified within the allowed time frame, it will be
deleted.

CSRF protection is disabled if `NODE_ENV` is set to 'development'. A warning
will be issued.