fix: re-expose devServer to local network
Due to a name clash between the `HOST` variable used in our `.env` and
the one used by webpack, the devServer wasn't getting exposed to LAN
like it would be expected. Change our variable to `HOSTNAME` to avoid
the clash and restore functionality.

Still relative to LAN access but too small for their own commit:

1. Add a new IP altname containing a sample local IP (to be modified by user).
2. Add a prompt for reminding the user to add said local IP.
3. Add `-extensions v3_req` option to openssl so that altnames are
   effectiverly added to the certificate.